+++ stable/xdhutl.h
@@ -133,9 +133,10 @@ namespace xdhutl {
 	E_AUTO( event_abstracts );
 
 	inline void GetTagDefaultEvent(
-		const str::string_ &Name,	// For the 'INPUT' tag, 'Name' contains the value of the 'type' attribute (uppercased).
+		const str::string_ &Tag,	// For the 'INPUT' tag, 'Name' contains the value of the 'type' attribute (uppercased).
 		str::string_ &Event,
-		str::dString &Keys )	// If the event is key related.
+		str::dString &Key );	// If the event is key related.
+#if 0
 	{
 		if ( Name == "SELECT" )
 			Event.Append( "change" );
@@ -156,6 +157,7 @@ namespace xdhutl {
 		}  else
 			Event.Append( "click" );
 	}
+#endif
 
 	void FillEventAbstract(
 		const str::string_ &DefaultEvent,